requirements.txt
discord.py==2.5.0
discord.py[voice]==2.5.0
SpeechRecognition==3.10.0
numba==0.57.0rc1
openai-whisper==20230314
openai==0.27.4
pydub==0.25.1
soundfile==0.12.1
python-dotenv==1.0.0
asyncio==3.4.3